Move some tests arounds
