+++ frontends/benchmarks/full-imperative/valid/SimpleArray.scala
